% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_default_occurrence_fields.R
\name{ala_default_occurrence_fields}
\alias{ala_default_occurrence_fields}
\title{Default Atlas occurrence fields}
\usage{
ala_default_occurrence_fields()
}
\value{
Character vector of field names.
}
\description{
Provides a reusable set of field names requested from the Atlas of Living
Australia when downloading occurrence records.
}
