% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cfg_read}
\alias{cfg_read}
\title{Read project configuration}
\usage{
cfg_read(profile = Sys.getenv("R_CONFIG_ACTIVE", "default"))
}
\arguments{
\item{profile}{Character scalar. Config profile name.}
}
\value{
A named list with configuration values.
}
\description{
Reads \code{config/config.yaml} using the active config profile (defaults to
\code{Sys.getenv("R_CONFIG_ACTIVE", "default")}). Returns a named list.
}
